requirements.txt
Django==1.10.5
Pygments==2.2.0
Werkzeug==0.11.15
appdirs==1.4.0
decorator==4.0.11
django-extensions==1.7.6
django-test-without-migrations==0.4
ipython==5.2.2
ipython-genutils==0.1.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyparsing==2.1.10
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
